chore: move sign use cases to accounts paste
