+++ src/app/useCases/accounts/SignInUseCase.ts
@@ -1,8 +1,8 @@
 import { compare } from 'bcryptjs';
 import { sign } from 'jsonwebtoken';
-import { env } from '../config/env';
-import { InvalidCredentials } from '../errors/InvalidCredentials';
-import { prismaClient } from '../lib/prismaClient';
+import { env } from '../../config/env';
+import { InvalidCredentials } from '../../errors/InvalidCredentials';
+import { prismaClient } from '../../lib/prismaClient';
 
 interface IInput {
   email: string;
